move ToggleControl to clients module
